Replace block users route checks with entity checks
In https://github.com/goalgorilla/open_social/pull/1903 we previously
introduced route level acces checks to hide blocked users. However, we
should implement this on the entity level so that it applies anywhere
user entities are used. This allows us to simplify our route permissions
and makes the chance of missing a spot less likely.