Implement address subfield management
This migrates the address sub-field settings from social_profile_fields
to the social_profile module in the new configuration form. While doing
so we move from custom settings to changing the field config (since the
address module actually already supports this).

While doing so we also update the default install config to use the
`fields_override` settings rather than the `fields` setting. This was a
change made in the `address` module which didn't update its config
itself but instead just pulled from `fields` if it was still used (and
only updated when the address config was actually updated through the
UI).

We still don't show the addressLine2 field since it doesn't have a label
(using the label from addressLine1) and showing it without addressLine1
doesn't make sense.

We override all the unmanageable fields to `hidden` and leave the
manageable sub-fields without override. This causes them to follow the
behaviour (of being "required" or "optional") of the main address field.

By setting a sub-field to "optional" it will be optional even if the
address field itself is required.

By setting a sub-field to "hidden" it will not be shown even if it would
normally be shown for the country's address format.

By setting a sub-field to "required" it will be shown and required
regardless of the selected country and the address field's field
settings.