Remove social_profile specific config from social_user
The admin view had a hidden dependency on social_profile. This was
originally added in 9a1dc5e1. The problem is that for other installer
related issues we now actually declare the dependency from
`social_profile` on `social_user`. This causes the first name and last
name fields not to exist when `social_user` is installed.

To solve this we adjust the user view so that it only contains things
that are included in `social_user`. Any profile specific modifications
are now made in the `social_profile` install hook.