Clean up previously completed TODOs
The user configured visibility will be ignored if the user is not
allowed to edit the value. This was implemented in 655b19.

The install config was implemented in 78ef84.

The nickname and nationality fields have configuration that matches
existing social_profile_fields functionality with the alter hook no
longer needed.

Third party settings were implemented in 78ef84.

The visibility field label configuration was fixed so that it matches
other field labels (capitalised first word, lowercase rest) and no
longer contains the placeholder styling.

Created https://www.drupal.org/project/social/issues/3250308 to handle
the query alter for active users.

SocialProfileFieldsOverride was removed in a811636 and the replacement
configuration was implemented in a commit before then.

An issue exists for the default address value:
https://www.drupal.org/project/social/issues/3238713

The SocialProfileFieldsSettingsForm is removed (with its route) since
the settings it controlled have been migrated entirely to the
SocialProfileSettingsForm.